Allow multiple test cases in the TShock test suite
Previously, we were initializing TShock before each test. This isn't
actually what we want to do, because NUnit will (by default) execute all
tests within one instance. Initializing TShock multiple times would
cause issues, so let's just do it once at the very beginning, before any
tests execute.
